% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_extract.R
\name{av_extract}
\alias{av_extract}
\alias{av_extract_df}
\alias{av_extract_fx}
\alias{av_narrowopts}
\title{Extract data from Alpha Vantage retuned data}
\usage{
av_extract_df(indta, grepstring = "", melt = F)

av_extract_fx(indta)

av_narrowopts(
  indta,
  expirationcode = "FW",
  mindays = 5,
  deltarange = c(0.1, 0.3)
)
}
\arguments{
\item{indta}{A data.table as returned by av_get()}

\item{grepstring}{select which variable (data item) to unnest in data.table returned from av_get_pf}

\item{melt}{Return data in melted/normalized form}

\item{expirationcode}{String of form \verb{[F(front)|B(ack)|A(ll)][W(eek)|M(onth)]}}

\item{mindays}{default 5. Minimum number of days to expiration to be considered a front contract}

\item{deltarange}{Narrow options to specific delta range}
}
\value{
Extracted data.table, summarized data,
}
\description{
Extract data from Alpha Vantage retuned data
}
\details{
\code{\link[=av_get_pf]{av_get_pf()}} frequently returns a nested data.table, or a structure with nested data.frames.  These are utilities functions to extract, filter and summarize returned values.
}
\examples{
\dontrun{
av_get_pf("","MARKET_STATUS")  |> av_extract_df()
av_get_pf("","TOP_GAINERS_LOSERS") |> av_extract_df("top_losers")
av_get_pf("USD/BRL","CURRENCY_EXCHANGE_RATE") |> av_extract_fx()
#' }

}
\seealso{
\code{\link[=av_get_pf]{av_get_pf()}}
}
