% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constants.R
\name{av_make_funcmap}
\alias{av_make_funcmap}
\title{Extract data from Alpha Vantage retuned data}
\usage{
av_make_funcmap()
}
\value{
function map dat.table, also written to data directory
}
\description{
Extract data from Alpha Vantage retuned data
}
\details{
Creates a data.table with API signatures and default values for Alphavantage API calls
}
\examples{
\dontrun{
av_make_funcmap()
}

}
\seealso{
\code{\link[=av_get_pf]{av_get_pf()}}
}
