% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constants.R
\name{av_reset_defaults}
\alias{av_reset_defaults}
\title{Extract data from Alpha Vantage retuned data}
\usage{
av_reset_defaults(fileopts = TRUE, keep_apikeys = FALSE, resetgrep = "*")
}
\arguments{
\item{fileopts}{(default: TRUE)  If TRUE, then remove all files and subdirectories}

\item{keep_apikeys}{(default: FALSE) Keep whatever API keys are stored}

\item{resetgrep}{(default "*") Only reset grepped default variables}
}
\value{
No return
}
\description{
Extract data from Alpha Vantage retuned data
}
\details{
Resets \code{\link[=av_runShiny]{av_runShiny()}} defaults to original (newly installed) state
}
\examples{
\dontrun{
av_reset_defaults()
}

}
\seealso{
\code{\link[=av_runShiny]{av_runShiny()}}
}
