% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_appRfuncs.R
\name{av_runShiny_interface}
\alias{av_runShiny_interface}
\alias{quick_message}
\alias{avsh_clipboard}
\alias{avsh_set_tabtitle}
\title{Helper functions for analytics}
\usage{
quick_message(wh, this_message = "", eval = TRUE, color = "#1f78b4")

avsh_clipboard(x, title = "")

avsh_set_tabtitle(newtext = "DETAIL", tabnm = "detail", makefocus = TRUE)
}
\arguments{
\item{wh}{inputID for shiny element to put a message underneath of.  See Documentation and/or Code}

\item{this_message}{(default "")  A text message to  be used. If empty string, the current message is cleared.}

\item{eval}{(default TRUE) OPtional parameter to suppress execution.}

\item{color}{Optional text color}

\item{x}{A \code{data.frame} or equivalent.}

\item{title}{String to add to a message printed if relevant}

\item{newtext}{(default"DETAIL") What to name the tab as}

\item{tabnm}{(default "detail") inputID of relevant tab}

\item{makefocus}{(default: TRUE) Upon setting the tab title, select the tab.}
}
\value{
logical value of \code{eval}

Nothing

Nothing
}
\description{
Displays a message underneath an input box

Copies a data.frame to the clipboard, with a status message if relevant

Sets the title for the Details tab
}
