% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_appRfuncs.R
\name{av_runShiny_userdata}
\alias{av_runShiny_userdata}
\alias{av_add_px}
\alias{av_add_earn}
\alias{av_load_shinydata}
\alias{av_add_assetgroups}
\alias{av_add_analytic}
\title{App database functions: Price}
\usage{
av_add_px(
  indta = NULL,
  assettypes = NULL,
  equitylist = NULL,
  dtstr = "-30y::",
  delay = 0
)

av_add_earn(
  substitute_earn = NULL,
  substitute_earnest = NULL,
  equitylist = NULL,
  delay = 0
)

av_load_shinydata(item = NULL)

av_add_assetgroups(indta)

av_add_analytic(runcode, func_name, helpstr = "user function", focus = "MAIN")
}
\arguments{
\item{indta}{A data.frame with two columns \code{c("listnm","ticker")} with one or more lines for each \code{"listnm"}}

\item{assettypes}{(default NULL)  An optional data.frame with minimal columns \code{c(symbol,type,currency,name)} with
descriptive data for the assets given in \code{indta}.  If not specified, a call to \code{av_get_pf(.,"SYMBOL_SEARCH")}
is necessary to determine the asset type (one of \code{c("Equity","ETF","FX","Index","Crypto")}) for subsequent
calls to \code{\link[=av_get_pf]{av_get_pf()}}}

\item{equitylist}{(default NULL) If specified, function will get equity prices from \code{av_get_pf}.  \code{indta} can be
null or is otherwise ignored.}

\item{dtstr}{(default \code{"-30y::"}). Date range to download if applicable.}

\item{delay}{(default 0) Seconds to delay calls to determine asset type for future AV downloads. This is
unused if \code{assettypes} is given.}

\item{substitute_earn}{A (default NULL)  data.frame with past earnings}

\item{substitute_earnest}{(default NULL)  A data.frame with  earnings estimates}

\item{item}{Any data name as seen by running \code{\link[=dump_state]{dump_state()}}.  \strong{If blank, loads entire database}}

\item{runcode}{Code string user must run to call the function.}

\item{func_name}{Name of function run when analytic is called.  If an empty string is supplied, the runcode will be de-registered.}

\item{helpstr}{(default: "user function"): A string comment to ad to the av.h (help) command}

\item{focus}{(default: "MAIN")  String with tab name to set focus to when command is run}
}
\value{
Nothing

Data.table with summary of downloaded or added earnings

Data item specified by \code{item} or a nothing (but a message) if left blank

Nothing

String message with success or failure of function addition.
}
\description{
Adds price data to \code{\link[=av_runShiny]{av_runShiny()}} internal data.

Adds earnings data to \code{\link[=av_runShiny]{av_runShiny()}} internal data, either by download or with user data

Loads internal data (prices, earnings, etc.

Adds asset lists to \code{\link[=av_runShiny]{av_runShiny()}} internal data.

Adds a user-defined function to the av Shiny app
}
\details{
Entire set of columns from \code{\link[=av_get_pf]{av_get_pf()}} can be added. First date column renamed to \code{timestamp} internally.

Entire set of columns from \code{\link[=av_get_pf]{av_get_pf()}} can be added. First date column renamed to \code{timestamp}.
If just assetypes is given, the function downloads earnings as needed (respecting maximum age parameters defined
in the app's \code{AVOPTS} tab.)
\strong{Note that price data must always be added first}

Lists are specified in normalized form.  Duplicate list names with those currently in use are replaced.

When the \code{\link[=av_runShiny]{av_runShiny()}} app is run, users can call functions to provide analytics based on asset strings in the command line.
This function allows users to add their own analytics by registering a function which takes, as inputs
\enumerate{
\item \code{todo}: The command line and any subsequent parameters as a space delimited string
\item \code{rv}: Reactive values supplied by the Shiny app.  In particular the parameter \code{rv$istr1} contains the semicolon delimited set of assets
prior to the command invocation.
The registered function should return a (possibly named, see vignette)
list containing one or more \code{gt()} tables, \code{dygraphs()}, or \code{ggplots()} to be displayed when the command is run.  See vignette for specfic details
}
\itemize{
\item The function specified must be available (i.e in \code{.GlobalENv()}) to the Shiny app when the command is run.  Otherwise an error message will be displayed.
\item If the specified command has already been registered, a message will be given and the internal data will be overridden.
}
}
\examples{
\dontrun{
# To add known symbols outside the app
av_add_px(equitylist=c("IBM","GS","JPM"))

# To add ad-hoc data from Alphavantage (e.g. Natgas spot at Henry Hub)
# Note that "symbol" in indta must match same in assettypes
asset_df <- data.frame(symbol=c("GAS_HH"),type=c("user"),currency=c("USD"), name=c("GAS_HH"))
ng_data <- av_get_pf("","NATURAL_GAS")[,.(symbol="GAS_HH",timestamp,close=value)]
av_add_px(ng_data, assettypes=asset_df)

# To data from other sources
suppressMessages(require(quantmod))
ffdta <- as.data.table(quantmod::getSymbols("FEDFUNDS",src="FRED",auto.assign=FALSE))
ffdta <- ffdta[,.(DT_ENTRY=index,close=FEDFUNDS,adjusted_close=FEDFUNDS,symbol="FEDFUNDS")]
av_add_px(ffdta)
}

\dontrun{
# To add earnings for a set of tickers
av_add_earn(equitylist=data.table(symbol=c("IBM","GS"))
}


\dontrun{
newtickers <- c("QQQ","QQQE","NDX")
av_add_assetgroups(data.table(listnm=rep("nasdaq",length(newtickers)),ticker=newtickers))
# To remove an asset list, just use an empty string for the ticker
av_add_assetgroups(data.table(listnm=c("new"),ticker=c("")))
}

\dontrun{
my_testfunc <- function(todo,rv) {
  message("todO: ",todo," with asset string ",rv$istr1)
  n_to_return <- c(strsplit(todo," "),"3")[[2]] |> as.numeric()
  table1 <- head(mtcars,n_to_return) |> gt()
  table2 <- data.table(asset=strsplit(rv$istr1,";")) |> gt()
  plot1 <- ggplot(mtcars,aes(mpg,disp)) + geom_point()
  return(list(table1, table2, plot1))
}
av_add_analytic("TEST","my_testfunc",helpstr="a test func")
# From the app; run "QQQ;SPY test 5"
# From the app: run "av.h"
}

}
\seealso{
\code{\link[=av_runShiny]{av_runShiny()}}
}
