% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_appRfuncs.R
\name{av_runShiny_userthings}
\alias{av_runShiny_userthings}
\alias{av_add_px}
\alias{av_add_earn}
\alias{av_add_assetgroups}
\alias{av_add_analytic}
\alias{av_shinydata}
\title{App database functions: Price}
\usage{
av_add_px(indta, assettypes = NULL, delay = 0)

av_add_earn(
  substitute_earn = NULL,
  substitute_earnest = NULL,
  assettypes = NULL,
  delay = 0
)

av_add_assetgroups(indta)

av_add_analytic(
  runcode,
  func_name,
  helpstr = "user function",
  focus = "MAIN",
  category = "USER"
)

av_shinydata(item = NULL)
}
\arguments{
\item{indta}{A data.frame with two columns \code{c("listnm","ticker")} with one or more lines for each \code{"listnm"}}

\item{assettypes}{(default NULL)  An optional data.frame with minimal columns \code{c(symbol,type,currency,name)} with
descriptive data for the assets given in \code{indta}.  If not specified, a call to \code{av_get_pf(.,"SYMBOL_SEARCH")}
is necessary to determine the asset type (one of \code{c("Equity","ETF","FX","Index","Crypto")}) for subsequent
calls to \code{\link[=av_get_pf]{av_get_pf()}}
NOTE THAT assettypes cannot be NULL if substitute_earn and substitute_earnest are NULL.}

\item{delay}{(default 0) Seconds to delay calls to determine asset type for future AV downloads. This is
unused if \code{assettypes} is given.}

\item{substitute_earn}{A (default NULL)  data.frame with past earnings}

\item{substitute_earnest}{(default NULL)  A data.frame with  earnings estimates}

\item{runcode}{Code string user must run to call the function.}

\item{func_name}{Name of function run when analytic is called.  If an empty string is supplied, the runcode will be de-registered.}

\item{helpstr}{(default: "user function"): A string comment to ad to the av.h (help) command}

\item{focus}{(default: "MAIN")  String with tab name to set focus to when command is run}

\item{category}{(default "USER") A string with a category used to sort function when help is called.}

\item{item}{Any data name as seen by running \code{\link[=dump_state]{dump_state()}}.  \strong{If blank, loads database}}
}
\value{
Nothing

Data.table with downloaded or added earnings

Nothing

String message with success or failure of function addition.

Data item specified by \code{item} or a nothing (but a message) if left blank
}
\description{
Adds price data to \code{\link[=av_runShiny]{av_runShiny()}} internal data.

Adds earnings data to \code{\link[=av_runShiny]{av_runShiny()}} internal data, either by download or with user data

Adds asset lists to \code{\link[=av_runShiny]{av_runShiny()}} internal data.

Adds a user-defined function to the av Shiny app

Loads internal data (prices, earnings, etc.
}
\details{
Entire set of columns from \code{\link[=av_get_pf]{av_get_pf()}} can be added. First date column renamed to \code{timestamp}

Entire set of columns from \code{\link[=av_get_pf]{av_get_pf()}} can be added. First date column renamed to \code{timestamp}.
If just assetypes is given, the function downloads earnings as needed (respecting maximum age parameters defined
in the app's \code{AVOPTS} tab.)

Lists are specified in normalized form.  Duplicate list names with those currently in use are replaced.

When the \code{\link[=av_runShiny]{av_runShiny()}} app is run, users can call functions to provide analytics based on asset strings in the command line.
This function allows users to add their own analytics by registering a function which takes, as inputs
\enumerate{
\item \code{todo}: The command line and any subsequent parameters as a space delimited string
\item \code{rv}: Reactive values supplied by the Shiny app.  In particular the parameter \code{rv$istr1} contains the semicolon delimited set of assets
prior to the command invocation.
The registered function should return a (possibly named, see vignette)
list containing one or more \code{gt()} tables, \code{dygraphs()}, or \code{ggplots()} to be displayed when the command is run.  See vignette for specfic details
}
\itemize{
\item The function specified must be available (i.e in \code{.GlobalENv()}) to the Shiny app when the command is run.  Otherwise an error message will be displayed.
\item If the specified command has already been registered, a message will be given and the internal data will be overridden.
}
}
\examples{
\dontrun{
av_add_px(av_get_pf("IBM","TIME_SERIES_DAILY_ADJUSTED"))
asset_df <- data.frame(symbol=c("HYG"),type=c("ETF"),currency=c("USD"), name=c("HY ETF"))
av_add_px(av_get_pf("HYG","TIME_SERIES_DAILY_ADJUSTED"), assettypes=asset_df)

suppressMessages(require(quantmod))
ffdta <- as.data.table(quantmod::getSymbols("FEDFUNDS",src="FRED",auto.assign=FALSE))
ffdta <- ffdta[,.(DT_ENTRY=index,close=FEDFUNDS,adjusted_close=FEDFUNDS,symbol="FEDFUNDS")]
av_add_px(ffdta)
}

\dontrun{
# To add earnings for a set of tickers
av_add_earn(assettypes=data.table(symbol=c("AAPL","QQQ"))
}


\dontrun{
newtickers <- c("QQQ","QQQE","NDX")
av_add_assetgroups(data.table(listnm=rep("nasdaq",length(newtickers)),ticker=newtickers))
# To remove an asset list, just use an empty string for the ticker
av_add_assetgroups(data.table(listnm=c("new"),ticker=c("")))
}

\dontrun{
my_testfunc <- function(todo,rv) {
  message("todO: ",todo," with asset string ",rv$istr1)
  n_to_return <- c(strsplit(todo," "),"3")[[2]] |> as.numeric()
  table1 <- head(mtcars,n_to_return) |> gt()
  table2 <- data.table(asset=strsplit(rv$istr1,";")) |> gt()
  plot1 <- ggplot(mtcars,aes(mpg,disp)) + geom_point()
  return(list(table1, table2, plot1))
}
av_add_analytic("TEST","my_testfunc",helpstr="a test func")
# From the app; run "QQQ;SPY test 5"
# From the app: run "av.h"
}

}
\seealso{
\code{\link[=av_runShiny]{av_runShiny()}}
}
