% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_appRfuncs.R
\name{avsh_quick_message}
\alias{avsh_quick_message}
\title{avsh_quick_message}
\usage{
avsh_quick_message(this_message, eval = TRUE, color = "#1f78b4")
}
\arguments{
\item{this_message}{(default "")  A text message to  be used. If empty string, the current message is cleared.}

\item{eval}{(default TRUE) OPtional parameter to suppress execution.}

\item{color}{Optional text color}
}
\value{
logical value of \code{eval}
}
\description{
Displays a message underneath an input box
}
\details{
Display a user message in the av_runShiny app
}
