% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_appRfuncs.R
\name{avsh_set_tabtitle}
\alias{avsh_set_tabtitle}
\title{avsh_set_tabtitle}
\usage{
avsh_set_tabtitle(newtext = "DETAIL", tabnm = "detail", makefocus = TRUE)
}
\arguments{
\item{newtext}{(default"DETAIL") What to name the tab as}

\item{tabnm}{(default "detail") inputID of relevant tab}

\item{makefocus}{(default: TRUE) Upon setting the tab title, select the tab.}
}
\value{
Nothing
}
\description{
Sets the title for the Details tab
}
\details{
Set an av_runShiny Tab Title
}
