% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_appRfuncs.R
\name{av_state_interface}
\alias{av_state_interface}
\alias{dump_state}
\alias{inv_rv}
\alias{dump_inv}
\alias{dump_assetgroups}
\alias{dump_av_funcs}
\alias{dump_captured}
\title{Exported Internal  internal state functions}
\usage{
dump_state(typegrep = "*")

inv_rv(rv)

dump_inv(invgrep = "*")

dump_assetgroups()

dump_av_funcs()

dump_captured(todo = "byfunction")
}
\arguments{
\item{typegrep}{: Grep string for internal state parameters}

\item{rv}{: An isolated list of av_shiny parameters}

\item{invgrep}{: A regular expression string}

\item{todo}{: One of c("byfunction","pxhist",any av function name)}
}
\value{
data.table with desired data.
}
\description{
retrieves internal data state of \code{\link[=av_runShiny]{av_runShiny()}}
\code{dump_state(typegrep="*")}
\code{dump_inv(invgrep="*")}
\code{dump_assetgroups()}
\code{dump_captured(todo="byfunction")}
\code{av_shiny_px()}
}
\examples{
\dontrun{
`dump_state()`
`dump_inv()`
`dump_av_funcs()`
`dump_assetgroups()`
`dump_captured(todo="byfunction")`
`inv_rv(rv)`
}

}
\seealso{
\code{\link[=av_runShiny]{av_runShiny()}}
}
