% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_themes.R
\name{gt.basetheme}
\alias{gt.basetheme}
\title{App formatting functions}
\usage{
gt.basetheme(
  x,
  gtopts = "all",
  sizepct = 100,
  style = 4,
  digits = 2,
  interactive_mult = 1.14,
  seps = FALSE,
  na_format = "-",
  size = "",
  interactive = ""
)
}
\arguments{
\item{x}{Input data or \code{gt} table}

\item{gtopts}{Which elements to add}

\item{sizepct}{(default: 70) How big to make tables}

\item{style}{(default: 4) Style number (see \code{\link[gt:gt]{gt()}})}

\item{digits}{(default: 2) Number formatting}

\item{interactive_mult}{(default: 80/70) Multtiple of \code{sizepct} applies to interactive tables.}

\item{seps}{(default: FALSE) Number formatting thousands separators}

\item{na_format}{(default: "-") WHat to show for NAs}

\item{size}{(default: "") Size of text, either \code{""} or \code{"small"}}

\item{interactive}{(default: "") String with interative elements: contains one of \code{(all|filter|search)}}
}
\description{
Adds default elements to \code{\link[gt:gt]{gt()}} table
}
