% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{message_if}
\alias{message_if}
\title{av_runShiny utilities}
\usage{
message_if(reallydothis, ...)
}
\arguments{
\item{reallydothis}{Condition to be met}

\item{...}{Additional items to be passed to \code{message}}
}
\description{
Prints message if condition met
}
