% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_appRfuncs.R
\name{quick_message}
\alias{quick_message}
\title{----------------------- Exported Shiny Functions}
\usage{
quick_message(wh, this_message = "", eval = TRUE, color = "#1f78b4")
}
\arguments{
\item{wh}{inputID for shiny element to put a message underneath of.  See Documentation and/or Code}

\item{this_message}{(default "")  A text message to  be used. If empty string, the current message is cleared.}

\item{eval}{(default TRUE) OPtional parameter to suppress execution.}

\item{color}{Optional text color}
}
\value{
logical value of \code{eval}
}
\description{
Displays a message underneath an input box
}
