% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_appRfuncs.R
\name{av_add_analytic}
\alias{av_add_analytic}
\title{av_add_analytic}
\usage{
av_add_analytic(runcode, func_name, helpstr = "user function", focus = "MAIN")
}
\arguments{
\item{runcode}{Code string user must run to call the function.}

\item{func_name}{Name of function run when analytic is called.  \strong{If an empty string is supplied, the runcode will be de-registered.}}

\item{helpstr}{(default: "user function"): A string comment to ad to the av.h (help) command}

\item{focus}{(default: "MAIN")  String with tab name to set focus to when command is run}
}
\value{
String message with success or failure of function addition.
}
\description{
Adds a user-defined function to the av Shiny app
}
\details{
Adds or removes a new command to the av_runShiny app

When the \code{\link[=av_runShiny]{av_runShiny()}} app is run, users can call functions to provide analytics based on asset strings in the command line.
This function allows users to add their own analytics by registering a function which takes, as inputs
\enumerate{
\item \code{todo}: The command line and any subsequent parameters as a space delimited string
\item \code{rv}: Reactive values supplied by the Shiny app.  In particular the parameter \code{rv$istr1} contains the semicolon delimited set of assets
prior to the command invocation.
The registered function should return a (possibly named, see vignette)
list containing one or more \code{gt()} tables, \code{dygraphs()}, or \code{ggplots()} to be displayed when the command is run.  See vignette for specfic details
}
\itemize{
\item The function specified must be available (i.e in \code{.GlobalENv()}) to the Shiny app when the command is run.  Otherwise an error message will be displayed.
\item If the specified command has already been registered, a message will be given and the internal data will be overridden.
}
}
\examples{
\dontrun{
my_testfunc <- function(todo,rv) {
  message("todO: ",todo," with asset string ",rv$istr1)
  n_to_return <- c(strsplit(todo," "),"3")[[2]] |> as.numeric()
  table1 <- head(mtcars,n_to_return) |> gt()
  table2 <- data.table(asset=strsplit(rv$istr1,";")) |> gt()
  plot1 <- ggplot(mtcars,aes(mpg,disp)) + geom_point()
  return(list(table1, table2, plot1))
}
av_add_analytic("TEST","my_testfunc",helpstr="a test func")
# From the app; run "QQQ;SPY test 5"
# From the app: run "av.h"
}
}
\seealso{
\code{\link[=av_runShiny]{av_runShiny()}}
}
