% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_appRfuncs.R
\name{av_add_assetgroups}
\alias{av_add_assetgroups}
\title{av_add_assetgroups}
\usage{
av_add_assetgroups(indta)
}
\arguments{
\item{indta}{A data.frame with two columns \code{c("listnm","ticker")} with one or more lines for each \code{"listnm"}}
}
\value{
Nothing
}
\description{
Adds asset lists to \code{\link[=av_runShiny]{av_runShiny()}} internal data.
}
\details{
Lists are specified in normalized form.  Duplicate list names with those currently in use are replaced.
}
\examples{
\dontrun{
newtickers <- c("QQQ","QQQE","NDX")
av_add_assetgroups(data.table(listnm=rep("nasdaq",length(newtickers)),ticker=newtickers))
# To remove an asset list, just use an empty string for the ticker
av_add_assetgroups(data.table(listnm=c("new"),ticker=c("")))
}
}
\seealso{
\code{\link[=av_runShiny]{av_runShiny()}}
}
