% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_db.R
\name{av_add_data}
\alias{av_add_data}
\title{App database functions}
\usage{
av_add_data(indta)
}
\arguments{
\item{indta}{A data.frame with the following minimal columns: \code{c(symbol,date,close,adjusted_close)}}
}
\value{
Nothing
}
\description{
Add price data to \code{\link[=av_runShiny]{av_runShiny()}} internal data
}
\details{
Entire set of columns from \code{\link[=av_get_pf]{av_get_pf()}} can be added. First date column renamed to \code{date}
}
\examples{
\dontrun{
av_add_data(av_get_pf("IBM","TIME_SERIES_DAILY_ADJUSTED"))
}

}
\seealso{
\code{\link[=av_runShiny]{av_runShiny()}}
}
