% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_appRfuncs.R
\name{av_add_earn}
\alias{av_add_earn}
\title{Add or download Earnings Data}
\usage{
av_add_earn(
  substitute_earn = NULL,
  substitute_earnest = NULL,
  equitylist = NULL,
  delay = 0
)
}
\arguments{
\item{substitute_earn}{A (default NULL)  data.frame with past earnings}

\item{substitute_earnest}{(default NULL)  A data.frame with  earnings estimates}

\item{equitylist}{(default NULL)  A list with tickers for which to retrieve earnings (from AlphaVantage)}

\item{delay}{(default 0)  A numeric value specifying delay between Alphavantage calls (in seconds)}
}
\value{
Data.table with summary of downloaded or added earnings
}
\description{
Adds earnings data to \code{\link[=av_runShiny]{av_runShiny()}} internal data, either by download or with user data
}
\details{
Add Earnings Data

Entire set of columns from \code{\link[=av_get_pf]{av_get_pf()}} can be added. First date column renamed to \code{timestamp}.
If just assetypes is given, the function downloads earnings as needed (respecting maximum age parameters defined
in the app's \code{AVOPTS} tab.)
\strong{Note that price data must always be added first}
}
\examples{
\dontrun{
# To add earnings for a set of tickers
av_load_shinydata()  # Make sure most recent data is loaded
av_add_earn(equitylist=data.table(symbol=c("IBM","GS"))
}
}
\seealso{
\code{\link[=av_runShiny]{av_runShiny()}}
}
