% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_appRfuncs.R
\name{av_add_earn}
\alias{av_add_earn}
\title{App database functions: Earnings}
\usage{
av_add_earn(
  substitute_earn = NULL,
  substitute_earnest = NULL,
  assettypes = NULL,
  delay = 0,
  maxage = 0
)
}
\arguments{
\item{substitute_earn}{A (default NULL)  data.frame with past earnings}

\item{substitute_earnest}{(default NULL)  A data.frame with  earnings estimates}

\item{assettypes}{(default NULL)  An optional data.frame with minimal columns \code{c(symbol,type,currency,name)} with
descriptive data for the assets given in \code{indta}.  If not specified, a call to \code{av_get_pf(.,"SYMBOL_SEARCH")}
is necessary to determine the asset type (one of \code{c("Equity","ETF","FX","Index","Crypto")}) for subsequent
calls to \code{\link[=av_get_pf]{av_get_pf()}}
NOTE THAT assettypes cannot be NULL if substitute_earn and substitute_earnest are NULL.}

\item{delay}{(default 0) Seconds to delay calls to determine asset type for future AV downloads. This is
unused if \code{assettypes} is given.}

\item{maxage}{(default 0) Maxium age of data before downloaded from Alphavantage}
}
\value{
Data.table with downloaded or added earnings
}
\description{
Adds earnings data to \code{\link[=av_runShiny]{av_runShiny()}} internal data, either by download or with user data
}
\details{
Entire set of columns from \code{\link[=av_get_pf]{av_get_pf()}} can be added. First date column renamed to \code{timestamp}
}
\examples{
\dontrun{
# To add earnings for a set of tickers
av_add_earn(assettypes=data.table(symbol=c("AAPL","QQQ"))
}
}
\seealso{
\code{\link[=av_runShiny]{av_runShiny()}}
}
