% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_appRfuncs.R
\name{av_add_options}
\alias{av_add_options}
\title{av_add_options}
\usage{
av_add_options(
  todo,
  dtstr = "-1w::",
  symbols = NULL,
  freq = "d",
  replace_data = FALSE,
  baddates_limits = 3,
  verbosity = "basic",
  external_path = NULL
)
}
\arguments{
\item{todo}{What to do to add options.  Choices are in the table below\tabular{cl}{
   \code{todo} \tab Description \cr
   \code{update} \tab Download and add or update full option runs by symbol over periods in \code{dtstr} \cr
   \code{inventory} \tab Create/update inventory of options that have been downloaded \cr
   \code{reconstruct_inventory} \tab Reconstruct inventory of all options that have been downloaded \cr
   \code{iv} \tab Updates internal data.table (\code{eqopt_iv} of Implied term structures by symbol and date \cr
   \code{copy_external} \tab Copies identical directory structures from \code{external_path} to siny app internal data \cr
}}

\item{dtstr}{(default \code{"-1w::"})  How far back to load options. ONly data needed within timeframe is downloaded, unless \code{replace_data=TRUE}}

\item{symbols}{(default: NULL) List of symbols to download}

\item{freq}{(Default \code{"d"}) Granularity of data downloaded ("d" for daily, "w" for weekly, "m" for monthly)}

\item{replace_data}{(Default \code{FALSE}) Update or replace data}

\item{baddates_limits}{(default 3): Consecutive business days of null data beyond which options are assumed not to exist.}

\item{verbosity}{(default \code{"time,basic"}) What to display as data is added.}

\item{external_path}{(default NULL) Path from which to replace existing data}
}
\value{
nothing
}
\description{
Manages option database and creates vol surfaces
}
\details{
Manage option data

\strong{This function can take a long time to run}.  Full option quotes are stored by symbol in arrow/parquet format by symbol
day, and contract.
\itemize{
\item Unless \code{replace_data==TRUE}, only dates and symbols that need to be downloaded will.
\item Except the \verb{OS S} function, all analytics run in the app require options data to be downloaded using this function.
}
}
\seealso{
\code{\link[=av_runShiny]{av_runShiny()}}
}
