% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_appRfuncs.R
\name{av_add_px}
\alias{av_add_px}
\title{av_add_px}
\usage{
av_add_px(
  indta = NULL,
  assettypes = NULL,
  equitylist = NULL,
  dtstr = "-30y::",
  delay = 0
)
}
\arguments{
\item{indta}{A data.frame with the following minimal columns: \code{c(symbol,timestamp,close)}.
Other variables added could be \code{c(adjusted_close,open,high,low,volume,dividend_amount,split_coefficient)}
If \code{adjusted_close} is not in the dataset, it will be set to \code{close}}

\item{assettypes}{(default NULL)  An optional data.frame with minimal columns \code{c(symbol,type,currency,name)} with
descriptive data for the assets given in \code{indta}.  If not specified, a call to \code{av_get_pf(.,"SYMBOL_SEARCH")}
is necessary to determine the asset type (one of \code{c("Equity","ETF","FX","Index","Crypto")}) for subsequent
calls to \code{\link[=av_get_pf]{av_get_pf()}}}

\item{equitylist}{(default NULL) If specified, function will get equity prices from \code{av_get_pf}.  \code{indta} can be
null or is otherwise ignored.}

\item{dtstr}{(default \code{"-30y::"}). Date range to download if applicable.}

\item{delay}{(default 0) Seconds to delay calls to determine asset type for future AV downloads. This is
unused if \code{assettypes} is given.}
}
\value{
Nothing
}
\description{
Adds price data to \code{\link[=av_runShiny]{av_runShiny()}} internal data.
}
\details{
Entire set of columns from \code{\link[=av_get_pf]{av_get_pf()}} can be added. First date column renamed to \code{timestamp} internally.
}
\examples{
\dontrun{
# To add known symbols outside the app
av_add_px(equitylist=c("IBM","GS","JPM"))

# To add ad-hoc data from Alphavantage (e.g. Natgas spot at Henry Hub)
# Note that "symbol" in indta must match same in assettypes
asset_df <- data.frame(symbol=c("GAS_HH"),type=c("user"),currency=c("USD"), name=c("GAS_HH"))
ng_data <- av_get_pf("","NATURAL_GAS")[,.(symbol="GAS_HH",timestamp,close=value)]
av_add_px(ng_data, assettypes=asset_df)

# To data from other sources
suppressMessages(require(quantmod))
ffdta <- as.data.table(quantmod::getSymbols("FEDFUNDS",src="FRED",auto.assign=FALSE))
ffdta <- ffdta[,.(DT_ENTRY=index,close=FEDFUNDS,adjusted_close=FEDFUNDS,symbol="FEDFUNDS")]
av_add_px(ffdta)
}
}
\seealso{
\code{\link[=av_runShiny]{av_runShiny()}}
}
