% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_extract.R
\name{av_extract_df}
\alias{av_extract_df}
\alias{av_extract_fx}
\alias{av_extract_analytics}
\alias{av_extract_divs_or_splits}
\title{Extract data from AlphaVantage API results}
\usage{
av_extract_df(indta, grepstring = "", melt = FALSE, empty_dt_onerror = TRUE)

av_extract_fx(indta, outputform = "common", cols = "", empty_dt_onerror = TRUE)

av_extract_analytics(indta, separate_vars = FALSE, empty_dt_onerror = TRUE)

av_extract_divs_or_splits(indta, empty_dt_onerror = TRUE)
}
\arguments{
\item{indta}{A data.table as returned by av_get()}

\item{grepstring}{select which variable (data item) to unnest in data.table returned from av_get_pf}

\item{melt}{Return data in melted/normalized form}

\item{empty_dt_onerror}{(default : TRUE): Return gracefully an empty data.table if requested item is not present.}

\item{outputform}{(default : \code{common}, \code{av_extract_fx()} only):  Use common names from \code{REALTIME_BULK_QUOTES}}

\item{cols}{(default : all columns:  \code{av_extract_fx()} only):  String or List of columns to return}

\item{separate_vars}{(default : FALSE)  separate out multiple levels of variable names into new keys}
}
\value{
Extracted data.tables for nested data returned from \code{\link[=av_get_pf]{av_get_pf()}}, If \code{grepstring} is not specified, first nested table is returned. \code{\link[=av_extract_fx]{av_extract_fx()}} returns a shortened data.table with FX quotes.
}
\description{
\code{av_extract_df()} pulls out nested data.frames from mixed data returned by  \code{\link[=av_get_pf]{av_get_pf()}}
\code{av_extract_fx()} returns a simplified FX quote in data.table formfrom \code{\link[=av_get_pf]{av_get_pf()}} calls.
\code{av_extract_analytics()} returns melted data.table from calls to \code{av_get_pf("ANALYTICS_FIXED_WINDOW")} or \code{av_get_pf("ANALYTICS_SLIDING_WINDOW")}
\code{av_extract_divs_or_splits()} returns melted data.table from calls to \code{av_get_pf("DIVIDENDS")} or \code{av_get_pf("SPLITS")}
}
\details{
\code{\link[=av_get_pf]{av_get_pf()}} frequently returns a nested data.table, or a structure with nested data.frames.  These are utilities functions to extract, filter and summarize returned values.
if \code{\link[=av_get_pf]{av_get_pf()}} returns a valid response, but empty extracted \code{data.tables}, am empty \code{data.table()} will be returned
NOTE: (Change 20260724:  Timestamp is returned in POSIXct form in local \code{Sys.timezone()} time)
}
\examples{
\dontrun{
av_get_pf("","MARKET_STATUS")  |> av_extract_df()
av_get_pf("","TOP_GAINERS_LOSERS") |> av_extract_df("top_losers")
av_get_pf("USD/BRL","CURRENCY_EXCHANGE_RATE") |> av_extract_fx()
av_get_pf(c("ORCL","IBM"),"ANALYTICS_FIXED_WINDOW") |> av_extract_analytics(separate_vars=TRUE)
}

}
\seealso{
\code{\link[=av_get_pf]{av_get_pf()}}, \code{\link[=av_grep_opts]{av_grep_opts()}}
}
