% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_get_pf.R
\name{av_get_pf}
\alias{av_get_pf}
\title{Get financial data from the Alpha Vantage API}
\usage{
av_get_pf(
  symbol,
  av_fun,
  symbolvarnm = "symbol",
  dfonerror = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{symbol}{A character string of an appropriate stock, fund, or currency
See parameter "symbol" in \href{https://www.alphavantage.co/documentation/}{Alpha Vantage API documentation}.}

\item{av_fun}{A character string matching an appropriate Alpha Vantage "function".
See parameter "function" in \href{https://www.alphavantage.co/documentation/}{Alpha Vantage API documentation}.}

\item{symbolvarnm}{Variable name of symbol requested inn output, default "symbol"}

\item{dfonerror}{Return an empty data.table when any error occurs, default TRUE}

\item{verbose}{Print debug information helpful for errors}

\item{...}{Additional parameters or overrides passed to the Alpha Vantage API.
For a list of parameters, visit the \href{https://www.alphavantage.co/documentation/}{Alpha Vantage API documentation}.}
}
\value{
Returns a data.table with results dependent on the function called.
Mixed data is returned as a melted data.table, possibly with nested data.frames.  Time series are returned as data.tables.
}
\description{
Get financial data from the Alpha Vantage API
}
\details{
\strong{The \code{av_fun} argument replaces the API parameter “function”} because function is a reserved name in R. All other arguments match the Alpha Vantage API parameters.

\strong{There is no need to specify the \code{apikey}, \code{datatype}, or \code{outputsize} parameters} as arguments to av_get_pf(). Before using, you must set the API key using av_api_key("YOUR_API_KEY"). \code{outputsize} defaults to "full" unless overridden with "compact in \code{...}."

\strong{Required parameters other than \code{symbol}} must be passed as named arguments via \code{...}.

\strong{Optional parameters have defaults} which can be obtained  by calling  \code{\link[=av_funhelp]{av_funhelp()}} and overridden via \code{...}.

\strong{There is no need to specify the datatype parameter} as an argument to \code{\link[=av_get_pf]{av_get_pf()}}. The function will return a data.table.

\strong{ForEx "FROM/TO" symbol details.} FOREX symbols in the \code{av_get_pf()} function are
supplied in \code{"FROM/TO"} format, which are then parsed in the Alpha Vantage API
into \code{from_currency} and \code{to_currency} API parameters. Usage example:
\code{av_get_pf(symbol = "EUR/USD", av_fun = "FX_DAILY")}
}
\examples{
\dontrun{

# SETUP API KEY
av_api_key("YOUR_API_KEY")

# ---- 1.0 STOCK TIME SERIES ----

# 1.1 TIME SERIES INTRADAY
av_get_pf("MSFT", av_fun = "TIME_SERIES_INTRADAY", interval = "5min", outputsize = "full")

# 1.2 TIME SERIES DAILY ADJUSTED
av_get_pf("MSFT", av_fun = "TIME_SERIES_DAILY_ADJUSTED", outputsize = "full")

# 1.3 QUOTE ENDPOINTS
av_get_pf("MSFT", av_fun = "GLOBAL_QUOTE")

# ---- 2.0 FOREX ----

# 2.1 CURRENCY EXCHANGE RATES
av_get_pf("EUR/USD", av_fun = "CURRENCY_EXCHANGE_RATE")

# 2.2 FX INTRADAY
av_get_pf("EUR/USD", av_fun = "FX_INTRADAY", interval = "5min", outputsize = "full")

# 2.3. FX DAILY
av_get_pf("EUR/USD", av_fun = "FX_DAILY", outputsize = "full")

# ---- 3.0 TECHNICAL INDICATORS ----

# 3.1 SMA
av_get_pf("MSFT", av_fun = "SMA", interval = "weekly", time_period = 10, series_type = "open")

# ---- 4.0 SECTOR PERFORMANCE ----

# 4.1 Sector Performance
av_get_pf(av_fun = "SECTOR")
}



}
\seealso{
\code{\link[=av_api_key]{av_api_key()}}, \code{\link[=av_extract_df]{av_extract_df()}}, \code{\link[=av_extract_fx]{av_extract_fx()}}, \code{\link[=av_narrowopts]{av_narrowopts()}},\code{\link[=av_funhelp]{av_funhelp()}}
}
