% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_extract_opts.R
\name{av_grep_opts}
\alias{av_grep_opts}
\title{Filter options from returned AV data}
\usage{
av_grep_opts(
  indta,
  grepstring = "F,M,C,otm",
  spot = NULL,
  mindays = 3,
  startdt = Sys.Date(),
  mindelta = 0.05,
  dropsymbol = FALSE
)
}
\arguments{
\item{indta}{A data.table as returned by \code{\link[=av_get_pf]{av_get_pf()}}}

\item{grepstring}{(default \verb{F,M,call,otm,act}) Three to five item string to select specific maturities, option types,  strike ranges, and
open interest categories from option sets returned by \code{av_get_pf(.,"HISTORICAL_OPTIONS")}. Each item in the list is an
abbreviated code for what to select. The items are not case sensitive and in order are\tabular{lll}{
   Item \tab Values \tab description \cr
   Expiration Limit \tab \code{F}|\code{B}|\code{A} \tab \code{F} is first maturity, \code{B} is second maturity, \code{A} (pr blank) are all \cr
   Maturity Type \tab \code{M}|\code{Q}|\code{A} \tab \code{M} for monthly contracts, \code{Q} for quarterly,  \code{A} (pr blank) are all \cr
   Option Type \tab \code{C}|\code{P}|\code{A} \tab Calls, puts, \code{all} (or blank) for all \cr
   Moneyness \tab \code{otm}|\code{itm}|\code{A} \tab Out of the money, in the money or both \cr
   Activity \tab \code{act}|\code{A} \tab If \code{A}, only pass contracts with positive open interest \cr
}}

\item{spot}{(default \code{NULL}) Spot to be used to determine itm/otm, If null then it is inferred from most out of the money call or put
Note: This parameter only applies if there is one symbol in \code{indta}.  If there is more than one ticker in \code{indta} a column \code{spot}
must be in \code{indta} to get correct results.}

\item{mindays}{(default 3). Minimum number of days to expiration to be passed through from \code{startdt}}

\item{startdt}{(default \code{Sys.Date()}). Date from which expirations will be considered.}

\item{mindelta}{(default \code{0.05}) delta limit on both moneyness sides, i.e. pass only options with deltas in range c(\code{mindelta},1-\code{mindelta})}

\item{dropsymbol}{(default \code{FALSE}). Drop symbol from returned data table.}
}
\value{
A reduced set of options obtained from \code{\link[=av_get_pf]{av_get_pf()}} using Alphavantage \code{HISTORICAL_OPTIONS} function.
}
\description{
Filter options from returned AV data
}
\details{
Alphavantage doesn't always return quarterly options

\code{\link[=av_get_pf]{av_get_pf()}} returns a large list of options.  This function helps to narrow down the list by maturity and moneyness.
}
\examples{
\dontrun{
av_get_pf("IBM","HISTORICAL_OPTIONS") |> av_grep_opts("F,M,put",mindays=2)
}

}
\seealso{
\code{\link[=av_get_pf]{av_get_pf()}}
}
