% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_appRfuncs.R
\name{av_load_shinydata}
\alias{av_load_shinydata}
\title{Load av_runShiny() internal data.}
\usage{
av_load_shinydata(item = NULL, verbose = TRUE)
}
\arguments{
\item{item}{Any data name as seen by running \code{\link[=dump_state]{dump_state()}}.  \strong{If blank, loads entire database}}

\item{verbose}{(default TRUE) write a status message to console}
}
\value{
Data item specified by \code{item} or a nothing (but a message) if left blank
}
\description{
Loads internal data (prices, earnings, etc.
}
\details{
Load ShinyApp data from Cache
}
\seealso{
\code{\link[=av_runShiny]{av_runShiny()}}
}
