% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_appRfuncs.R
\name{av_load_shinydata}
\alias{av_load_shinydata}
\alias{av_shinydata}
\title{App database functions}
\usage{
av_shinydata(item = NULL)
}
\arguments{
\item{item}{Any data name as seen by running \code{\link[=dump_state]{dump_state()}}.  \strong{If blank, loads database}}
}
\value{
Data item specified by \code{item} or a nothing (but a message) if left blank
}
\description{
Loads internal data (prices, earnings, etc.
}
\seealso{
\code{\link[=av_runShiny]{av_runShiny()}}
}
