% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_extract_opts.R
\name{av_opt_helper_cols}
\alias{av_opt_helper_cols}
\title{Add additional data to returned option sets}
\usage{
av_opt_helper_cols(indta, scaling = NULL, spot = NULL)
}
\arguments{
\item{indta}{An option data.table as returned by \code{\link[=av_get_pf]{av_get_pf()}}}

\item{scaling}{(default \code{NULL}) Scaling factor for marks, last values, and greeks.  Options are
\itemize{
\item \code{NULL} or \code{"none"} : No Scaling
\item \code{"10contracts"} : 10 contracts converted into market value
\item \code{"10kMV"} : 10,000 USD converted into equivalent market value
\item \code{number} : Any numeric value in thousands of USD into equivalent market value
}}

\item{spot}{(default \code{NULL}) Spot to be used to determine itm/otm, If null then it is inferred from most out of the money call or put
Note: This parameter only applies if there is one symbol in \code{indta}.  If there is more than one ticker in \code{indta} a column \code{spot}
must be in \code{indta} to get correct results.}
}
\value{
An option \code{data.table} with extra columns helpful for further analysis
}
\description{
Add additional data to returned option sets
}
\details{
Adds columns including\tabular{cl}{
   Column \tab Definition \cr
   \code{daysExp} \tab Days to Expiration \cr
   \code{bo_pct} \tab Bid offer in percent of option mid \cr
   \code{bid_size_poi} \tab Bid size percent of open interest \cr
   \code{ncak} \tab Notional number of contracts \cr
}
}
\examples{
\dontrun{
av_get_pf("IBM","HISTORICAL_OPTIONS") |> av_grep_opts("F,M,put",mindays=2) |> av_opt_helper_cols()
}

}
\seealso{
\code{\link[=av_grep_opts]{av_grep_opts()}}
}
