% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constants.R
\name{av_reset_defaults}
\alias{av_reset_defaults}
\title{Extract data from Alpha Vantage retuned data}
\usage{
av_reset_defaults()
}
\value{
No return
}
\description{
Extract data from Alpha Vantage retuned data
}
\details{
Resets \code{\link[=av_runShiny]{av_runShiny()}} defaults to original (newly installed) state
}
\examples{
\dontrun{
av_reset_defaults()
}

}
\seealso{
\code{\link[=av_runShiny]{av_runShiny()}}
}
