% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_runapp.R
\name{av_runShiny}
\alias{av_runShiny}
\title{Set the Alpha Vantage API Key}
\usage{
av_runShiny()
}
\value{
Nothing
}
\description{
\code{av_runShiny()} sets Alphavantage API key and entitlement code
}
\details{
Starts Shiny APp
}
\examples{
\dontrun{
av_runShiny()
}

}
