% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_api_key.R
\name{avpf_set_request_pace}
\alias{avpf_set_request_pace}
\title{Set the Alpha Vantage max requests per minute}
\usage{
avpf_set_request_pace(max_requests_per_min)
}
\arguments{
\item{max_requests_per_min}{(default 60) Maximum requests per minute}
}
\value{
Nothing
}
\description{
\code{avpf_set_request_pace()} sets Alphavantage API key and entitlement code
}
\details{
Change this if you have a better plan with AlphaVantage
}
\seealso{
\code{\link[=avpf_api_key]{avpf_api_key()}}
}
