% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_appRfuncs.R
\name{avsh_clipboard}
\alias{avsh_clipboard}
\title{avsh_clipboard}
\usage{
avsh_clipboard(x, title = "")
}
\arguments{
\item{x}{A \code{data.frame} or equivalent.}

\item{title}{String to add to a message printed if relevant}
}
\value{
Nothing
}
\description{
Copies a data.frame to the clipboard, with a status message if relevant
}
\details{
Copy data to clipboard
}
