% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_db.R
\name{dump_the}
\alias{dump_the}
\alias{dump_inv}
\alias{dump_assetlist}
\alias{dump_captured}
\title{Extract internal state}
\usage{
dump_the(typegrep = "*")

dump_inv()

dump_assetlist(returngt = TRUE)

dump_captured(todo = "byfunction")
}
\arguments{
\item{typegrep}{: Grep string for internal state parameters}

\item{returngt}{: Return GT table}

\item{todo}{: One of c("byfunction","pxhist",any av function name)}
}
\value{
data.table with desired data.
}
\description{
Prints internal data state of \code{\link[=av_runShiny]{av_runShiny()}}
\code{dump_the(typegrep="*")}
\code{dump_inv()}
\code{dump_assetlist()}
\code{dump_captured()}
}
\examples{
\dontrun{
`dump_the()`
`dump_inv()`
`dump_assetlist(returngt=TRUE)`
`dump_captured(todo="byfunction")`
}

}
\seealso{
\code{\link[=av_runShiny]{av_runShiny()}}
}
