% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_themes.R
\name{gt.basetheme}
\alias{gt.basetheme}
\title{App formatting functions}
\usage{
gt.basetheme(
  x,
  gtopts = "all",
  sizepct = 70,
  style = 4,
  digits = 2,
  seps = FALSE,
  na_format = "-",
  interactive = FALSE
)
}
\arguments{
\item{x}{Input data or \code{gt} table}

\item{gtopts}{Which elements to add}

\item{sizepct}{(default: 70) How big to make tables}

\item{style}{(default: 4) Style number (see \code{\link[gt:gt]{gt::gt()}})}

\item{digits}{(default: 2) Number formatting}

\item{seps}{(default: FALSE) Number formatting thousands separators}

\item{na_format}{(default: "-") WHat to show for NAs}

\item{interactive}{(default: \code{FALSE}) Add interactive elements}
}
\description{
Adds default elements to \code{\link[gt:gt]{gt::gt()}} table
}
